making compatible with module
