--- gradle.properties
@@ -1,13 +1,0 @@
-#Gradle
-org.gradle.jvmargs=-Xmx2048M -Dkotlin.daemon.jvm.options\="-Xmx2048M"
-
-#Kotlin
-kotlin.code.style=official
-
-#Android
-android.useAndroidX=true
-
-#MPP
-kotlin.mpp.enableGranularSourceSetsMetadata=true
-kotlin.native.enableDependencyPropagation=false
-kotlin.mpp.enableCInteropCommonization=true
