--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Jul 09 23:51:06 IST 2022
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.3.3-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
