--- kmmLearning/src/androidTest/kotlin/com/neural/kmmlearning/androidTest.kt
@@ -1,12 +1,0 @@
-package com.neural.kmmlearning
-
-import org.junit.Assert.assertTrue
-import org.junit.Test
-
-class AndroidGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue("Check Android is mentioned", Greeting().greeting().contains("Android"))
-    }
-}
