--- kmmLearning/src/commonTest/kotlin/com/neural/kmmlearning/commonTest.kt
@@ -1,12 +1,0 @@
-package com.neural.kmmlearning
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class CommonGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greeting().contains("Hello"), "Check 'Hello' is mentioned")
-    }
-}
