--- kmmLearning/src/iosTest/kotlin/com/neural/kmmlearning/iosTest.kt
@@ -1,12 +1,0 @@
-package com.neural.kmmlearning
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class IosGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greeting().contains("iOS"), "Check iOS is mentioned")
-    }
-}
