--- settings.gradle.kts
@@ -1,10 +1,0 @@
-pluginManagement {
-    repositories {
-        google()
-        gradlePluginPortal()
-        mavenCentral()
-    }
-}
-
-rootProject.name = "kmmLearning"
-include(":kmmLearning")
