+++ src/androidMain/AndroidManifest.xml
