+++ src/commonMain/kotlin/com/neural/kmmlearning/HelloKMM.kt
@@ -1,0 +1,9 @@
+package com.neural.kmmlearning
+
+class HelloKMM {
+
+    fun hello(msg: String): String {
+        println(msg)
+        return "Hello from KMM"
+    }
+}
